Fix build issue by moving authOptions to separate file and changing nextauth handler
