+++ app/api/auth/[...nextauth]/route.ts
@@ -1,65 +1,7 @@
-import NextAuth, { AuthOptions } from "next-auth";
-import CredentialsProvider from "next-auth/providers/credentials";
-import { compare } from "bcrypt";
-import { sql } from "@vercel/postgres";
-
-export const authOptions: AuthOptions = {
-  session: {
-    strategy: "jwt",
-  },
-  pages: {
-    signIn: "/login",
-  },
-  providers: [
-    CredentialsProvider({
-      credentials: {
-        email: {},
-        password: {},
-      },
-      async authorize(credentials, req) {
-        const response = await sql`
-        SELECT * FROM users WHERE email=${credentials?.email}`;
-        const user = response.rows[0];
-
-        const passwordCorrect = await compare(
-          credentials?.password || "",
-          user.password
-        );
-
-        if (passwordCorrect) {
-          return {
-            id: user.id,
-            role: user.role,
-          };
-        }
-
-        return null;
-      },
-    }),
-  ],
-  callbacks: {
-    async jwt({ token, user, session }) {
-      console.log("jwt callback", { token, user, session });
-      if (user) {
-        return {
-          ...token,
-          role: user.role,
-        };
-      }
-      return token;
-    },
-    async session({ session, token, user }) {
-      return {
-        ...session,
-        user: {
-          ...session.user,
-          id: token.sub,
-          role: token.role,
-        },
-      };
-    },
-  },
-};
+// app/api/auth/[...nextauth]/route.ts
+import NextAuth from "next-auth";
+import { authOptions } from "../authOptions";
 
 const handler = NextAuth(authOptions);
+
 export { handler as GET, handler as POST };
